requirements.txt
Pillow
pycryptodome
